Allow you to pass a proc to push for resolving refs
The new refs_resolver param allows you to pass a proc that can do some additional ref processing within the write locks, but before the actual push.

We are making use of this when committing to the repo directly from rugged. When we call Mirror.push we know our target branch, but not the commit id yet, so within the locks, in this resolver, we will create the actual commit object.