Adjusting SSH to auto-trust new servers if they haven't been seen before.
I verified this is quite easy to do. We just pass StrictHostKeyChecking=no

According to both the docs and my own testing this will cause SSH to
automatically trust new servers but you'll still fail out with an
error (as we desire) if the server has been seen before with a different
fingerprint.