Fixing the exception message for 'info' errors against git-fusion.
The logic had previously been:
  fail RunError, output[/^fatal: (?<error>.*)$/]

I believe this is the third time I've had to remove this bad code :(

The result of this code is you get the unadjusted message if it contains a leading "fatal: "
(e.g. it doesn't strip the fatal: it just returns the full text)

If the text doesn't lead with "fatal: " you get nil.

This tends to mask useful messages and is inconsistent  with the behaviour of 'list'.

Fixing this allows us to better surface errors on the enable mirroring page.